list interactions for saved memory

+++ hdr/neighbours.hpp
@@ -33,8 +33,8 @@ namespace neighbours{
    class neighbour_t {
 	public:
 
-		int nn; // atom id of neighbour
-		int i; // interaction type of neighbour
+		uint64_t nn; // atom id of neighbour
+		uint64_t i; // interaction type of neighbour
 
       double vx; // real coordinate vector between atoms i->j
       double vy;
